websockets-server.lisp: expose idmap multithreading bug
test case: (call-in-ws-repl "alert('foo')")

- when the query times out after 5 seconds (the default)
the mbox object is removed from *id-map*.
During this time the websocket-driver:on :message callback handler
does not run.
by the time that runs and handle-js-query gets called
the mbox object is gone already.

the problem seems to be that blocking in websocket-driver:send
prevents other messages from being processed.